Move strategies to own package
